adds a retry for app instantiation
when running in a container, occasionally, there is a transient network error during app instantiation. this allows for retrying rather than just dying and having to completely reinstantiate a new container. obviously, if there is a bigger error, or the network problem is longer than a few seconds, this will fail after 5 retries and a whole new container will be launched.